Got the app working!
Refactored everything, set up input and file reading, so now the app works just about the same as it did in Eclipse.  The two main differences I see are the borders are not drawn yet, and I removed the GameController class, since the reorganization of LevelView made it somewhat redundant alongside LevelManager.