Added level indicator, changed level csv code, fixed margin bug, misc.
Now, when selecting a level, the current level will be indicated in lime
green.  Additionally, added a textView at the top of the level select
screen which indicates that it is levels you are selecting.  Made major
changes to the way csv files are read in and processed.  Now, instead
of a single character to indicate what goes where, two characters are
used per space, allowing gameObjects to be in the same place as
triggers.  Fixed the margin issue.  Made miscellaneous small changes.