Added a basic test set for player behaviors.  Made minor code changes.
Added a (so far incomplete) set of tests for player behaviors.  This
just basically creates a simple level, tries moves, and asserts that
the outcomes match the expected outcomes given the rules of the game.
This will allow a quick way to make sure that no changes to movement
break the game (e.g. make impossible levels after subtle functionality
changes).  Additionally, made some minor code changes, like changing
all of the += 1 in loops to ++, and adding an empty level constructor.