Added sound to level select, fixed undo sound to only play on change.
Added the click sound to the level select button, and to the buttons
on the level select activity.  Now, the click happens pretty much
with every click.  Additionally, made it so that the undo sound really
only plays when the undo action takes place.  Changed Level.processInput
to return whether a move actually occurs or not.