Fixed sound issue, added content descriptions. Finalized for release.
Removed the part in the toggleSound() method of the PlayActivity that
turned off sound.  This turned out to require permissions on some phones,
and there was no reason for it to be there.  Additionally, it may have
caused a phone to not ring out loud if the app was running.  Also added
content descriptions, performed some cleanup, got ready for the full
release.